Add Alexa Auth Handler.
