+++ libraries/Bot.Builder.Community.Adapters.Alexa/ValidationHelper.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 using Alexa.NET.Request;
+using Bot.Builder.Community.Adapters.Alexa.Core;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 
@@ -9,42 +10,11 @@ namespace Bot.Builder.Community.Adapters.Alexa
     internal class ValidationHelper
     {
         public static async Task<bool> ValidateRequest(HttpRequest request, SkillRequest skillRequest, string body, ILogger logger)
-        { 
-            request.Headers.TryGetValue("SignatureCertChainUrl", out var signatureChainUrl);
-            if (string.IsNullOrWhiteSpace(signatureChainUrl))
-            {
-                logger.LogError("Validation failed due to empty SignatureCertChainUrl header");
-                return false;
-            }
+        {
+            request.Headers.TryGetValue(AlexaMessageAuthorizationHandler.SignatureCertChainUrlHeader, out var signatureChainUrls);
+            request.Headers.TryGetValue(AlexaMessageAuthorizationHandler.SignatureHeader, out var signatureHeaders);
 
-            Uri certUrl;
-            try
-            {
-                certUrl = new Uri(signatureChainUrl);
-            }
-            catch
-            {
-                logger.LogError($"Validation failed. SignatureChainUrl not valid: {signatureChainUrl}");
-                return false;
-            }
-
-            request.Headers.TryGetValue("Signature", out var signature);
-            if (string.IsNullOrWhiteSpace(signature))
-            {
-                logger.LogError("Validation failed - Empty Signature header");
-                return false;
-            }
-
-            var isTimestampValid = RequestVerification.RequestTimestampWithinTolerance(skillRequest);
-            var valid = await RequestVerification.Verify(signature, certUrl, body);
-
-            if (!valid || !isTimestampValid)
-            {
-                logger.LogError("Validation failed - RequestVerification failed");
-                return false;
-            }
-
-            return true;
+            return await new AlexaMessageAuthorizationHandler(logger).ValidateSkillRequest(skillRequest, body, signatureChainUrls, signatureHeaders).ConfigureAwait(false);
         }
     }
 }
